+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added basic support for uv. `uv run pyodide` will now work.
+  [#132](https://github.com/pyodide/pyodide-build/pull/132)
+
 ### Changed
 
 - The Rust toolchain version has been updated to `nightly-2025-01-18`.